backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
clickhouse-driver==0.2.6
python-jose==3.3.0
python-multipart==0.0.6
pandas==2.1.3
pydantic==2.5.2
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0 